+++ src/message/data/u4.rs
@@ -3,9 +3,11 @@
 use crate::message::data::{FromClamped, FromOverFlow};
 
 /// A primitive value that can be from 0-0x0F
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub struct U4(u8);
 
 /// Error representing that this value is not a valid u4
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub struct InvalidU4(pub u8);
 
 impl TryFrom<u8> for U4 {