+++ src/message/mod.rs
@@ -6,23 +6,22 @@ pub mod data;
 pub mod notes;
 pub mod raw;
 
-use crate::message::channel::Channel;
-use crate::message::control_function::ControlFunction;
-use crate::message::data::u7::U7;
-use crate::message::data::FromClamped;
-use crate::message::notes::Note;
+pub use crate::message::channel::Channel;
+pub use crate::message::control_function::ControlFunction;
+pub use crate::message::data::u14::U14;
+pub use crate::message::data::u7::U7;
+pub use crate::message::data::{FromClamped, FromOverFlow};
+pub use crate::message::notes::Note;
+
 use crate::message::raw::{Payload, Raw};
 use crate::packet::cable_number::CableNumber;
 use crate::packet::code_index_number::CodeIndexNumber;
 use crate::packet::{UsbMidiEventPacket, UsbMidiEventPacketError};
 
-type Velocity = U7;
+/// Note velocity value.
+pub type Velocity = U7;
 
-/// Represents midi messages.
-///
-/// Note: not current exhaustive and SysEx messages end up
-/// being a confusing case. So are currently note implemented
-/// they are sort-of unbounded
+/// Represents the MIDI messages.
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub enum Message {
     /// Note On message.
@@ -36,9 +35,31 @@ pub enum Message {
     /// Channel aftertouch message.
     ChannelAftertouch(Channel, U7),
     /// Pitchwheel message.
-    PitchWheelChange(Channel, U7, U7),
+    PitchWheelChange(Channel, U14),
     /// Control Change (CC) message.
     ControlChange(Channel, ControlFunction, U7),
+    /// MTC Quarter Frame message.
+    MtcQuarterFrame(U7),
+    /// Song Position Pointer message.
+    SongPositionPointer(U14),
+    /// Song Select message.
+    SongSelect(U7),
+    /// Tune Request message.
+    TuneRequest,
+    /// Timing Clock message.
+    TimingClock,
+    /// Tick message.
+    Tick,
+    /// Start message.
+    Start,
+    /// Continue message.
+    Continue,
+    /// Stop message.
+    Stop,
+    /// Active Sensing message.
+    ActiveSensing,
+    /// Reset message.
+    Reset,
 }
 
 const NOTE_OFF_MASK: u8 = 0b1000_0000;
@@ -48,6 +69,18 @@ const PROGRAM_MASK: u8 = 0b1100_0000;
 const CHANNEL_AFTERTOUCH_MASK: u8 = 0b1101_0000;
 const PITCH_BEND_MASK: u8 = 0b1110_0000;
 const CONTROL_CHANGE_MASK: u8 = 0b1011_0000;
+
+const MTC_QUARTER_FRAME: u8 = 0xF1;
+const SONG_POSITION_POINTER: u8 = 0xF2;
+const SONG_SELECT: u8 = 0xF3;
+const TUNE_REQUEST: u8 = 0xF6;
+const TIMING_CLOCK: u8 = 0xF8;
+const TICK: u8 = 0xF9;
+const START: u8 = 0xFA;
+const CONTINUE: u8 = 0xFB;
+const STOP: u8 = 0xFC;
+const ACTIVE_SENSING: u8 = 0xFE;
+const RESET: u8 = 0xFF;
 
 impl From<Message> for Raw {
     fn from(value: Message) -> Raw {
@@ -77,7 +110,10 @@ impl From<Message> for Raw {
                 let status = CHANNEL_AFTERTOUCH_MASK | u8::from(chan);
                 Raw { status, payload }
             }
-            Message::PitchWheelChange(chan, lsb, msb) => {
+            Message::PitchWheelChange(chan, value) => {
+                let value = u16::from(value);
+                let lsb = U7::try_from((value & 0x7F) as u8).unwrap();
+                let msb = U7::try_from(((value >> 7) & 0x7F) as u8).unwrap();
                 let payload = Payload::DoubleByte(lsb, msb);
                 let status = PITCH_BEND_MASK | u8::from(chan);
                 Raw { status, payload }
@@ -85,6 +121,64 @@ impl From<Message> for Raw {
             Message::ControlChange(chan, control_function, value) => {
                 let payload = Payload::DoubleByte(control_function.0, value);
                 let status = CONTROL_CHANGE_MASK | u8::from(chan);
+                Raw { status, payload }
+            }
+            Message::MtcQuarterFrame(frame) => {
+                let payload = Payload::SingleByte(frame);
+                let status = MTC_QUARTER_FRAME;
+                Raw { status, payload }
+            }
+            Message::SongPositionPointer(value) => {
+                let value = u16::from(value);
+                let lsb = U7::try_from((value & 0x7F) as u8).unwrap();
+                let msb = U7::try_from(((value >> 7) & 0x7F) as u8).unwrap();
+                let payload = Payload::DoubleByte(lsb, msb);
+                let status = SONG_POSITION_POINTER;
+                Raw { status, payload }
+            }
+            Message::SongSelect(song) => {
+                let payload = Payload::SingleByte(song);
+                let status = SONG_SELECT;
+                Raw { status, payload }
+            }
+            Message::TuneRequest => {
+                let payload = Payload::Empty;
+                let status = TUNE_REQUEST;
+                Raw { status, payload }
+            }
+            Message::TimingClock => {
+                let payload = Payload::Empty;
+                let status = TIMING_CLOCK;
+                Raw { status, payload }
+            }
+            Message::Tick => {
+                let payload = Payload::Empty;
+                let status = TICK;
+                Raw { status, payload }
+            }
+            Message::Start => {
+                let payload = Payload::Empty;
+                let status = START;
+                Raw { status, payload }
+            }
+            Message::Continue => {
+                let payload = Payload::Empty;
+                let status = CONTINUE;
+                Raw { status, payload }
+            }
+            Message::Stop => {
+                let payload = Payload::Empty;
+                let status = STOP;
+                Raw { status, payload }
+            }
+            Message::ActiveSensing => {
+                let payload = Payload::Empty;
+                let status = ACTIVE_SENSING;
+                Raw { status, payload }
+            }
+            Message::Reset => {
+                let payload = Payload::Empty;
+                let status = RESET;
                 Raw { status, payload }
             }
         }
@@ -99,6 +193,43 @@ impl TryFrom<&[u8]> for Message {
             Some(byte) => byte,
             None => return Err(UsbMidiEventPacketError::MissingDataPacket),
         };
+
+        match *status_byte {
+            MTC_QUARTER_FRAME => {
+                return Ok(Message::MtcQuarterFrame(get_u7_at(data, 1)?));
+            }
+            SONG_POSITION_POINTER => {
+                return Ok(Message::SongPositionPointer(get_u14(data)?));
+            }
+            SONG_SELECT => {
+                return Ok(Message::SongSelect(get_u7_at(data, 1)?));
+            }
+            TUNE_REQUEST => {
+                return Ok(Message::TuneRequest);
+            }
+            TIMING_CLOCK => {
+                return Ok(Message::TimingClock);
+            }
+            TICK => {
+                return Ok(Message::Tick);
+            }
+            START => {
+                return Ok(Message::Start);
+            }
+            CONTINUE => {
+                return Ok(Message::Continue);
+            }
+            STOP => {
+                return Ok(Message::Stop);
+            }
+            ACTIVE_SENSING => {
+                return Ok(Message::ActiveSensing);
+            }
+            RESET => {
+                return Ok(Message::Reset);
+            }
+            _ => {}
+        }
 
         let event_type = status_byte & 0b1111_0000;
         let channel_bytes = (status_byte) & 0b0000_1111;
@@ -123,11 +254,7 @@ impl TryFrom<&[u8]> for Message {
             )),
             PROGRAM_MASK => Ok(Message::ProgramChange(channel, get_u7_at(data, 1)?)),
             CHANNEL_AFTERTOUCH_MASK => Ok(Message::ChannelAftertouch(channel, get_u7_at(data, 1)?)),
-            PITCH_BEND_MASK => Ok(Message::PitchWheelChange(
-                channel,
-                get_u7_at(data, 1)?,
-                get_u7_at(data, 2)?,
-            )),
+            PITCH_BEND_MASK => Ok(Message::PitchWheelChange(channel, get_u14(data)?)),
             CONTROL_CHANGE_MASK => Ok(Message::ControlChange(
                 channel,
                 ControlFunction(get_u7_at(data, 1)?),
@@ -180,10 +307,21 @@ impl Message {
             Self::NoteOn(_, _, _) => CodeIndexNumber::NoteOn,
             Self::NoteOff(_, _, _) => CodeIndexNumber::NoteOff,
             Self::ChannelAftertouch(_, _) => CodeIndexNumber::ChannelPressure,
-            Self::PitchWheelChange(_, _, _) => CodeIndexNumber::PitchBendChange,
+            Self::PitchWheelChange(_, _) => CodeIndexNumber::PitchBendChange,
             Self::PolyphonicAftertouch(_, _, _) => CodeIndexNumber::PolyKeyPress,
             Self::ProgramChange(_, _) => CodeIndexNumber::ProgramChange,
             Self::ControlChange(_, _, _) => CodeIndexNumber::ControlChange,
+            Self::MtcQuarterFrame(_) => CodeIndexNumber::SystemCommon2Bytes,
+            Self::SongPositionPointer(_) => CodeIndexNumber::SystemCommon3Bytes,
+            Self::SongSelect(_) => CodeIndexNumber::SystemCommon2Bytes,
+            Self::TuneRequest => CodeIndexNumber::SystemCommon1Byte,
+            Self::TimingClock => CodeIndexNumber::SingleByte,
+            Self::Tick => CodeIndexNumber::SingleByte,
+            Self::Start => CodeIndexNumber::SingleByte,
+            Self::Continue => CodeIndexNumber::SingleByte,
+            Self::Stop => CodeIndexNumber::SingleByte,
+            Self::ActiveSensing => CodeIndexNumber::SingleByte,
+            Self::Reset => CodeIndexNumber::SingleByte,
         }
     }
 }
@@ -201,6 +339,12 @@ fn get_u7_at(data: &[u8], index: usize) -> Result<U7, UsbMidiEventPacketError> {
     Ok(U7::from_clamped(data_byte))
 }
 
+fn get_u14(data: &[u8]) -> Result<U14, UsbMidiEventPacketError> {
+    let lsb = get_byte_at_position(data, 1)?;
+    let msb = get_byte_at_position(data, 2)?;
+    Ok(U14::from_clamped(((msb as u16) << 7) | (lsb as u16)))
+}
+
 fn get_byte_at_position(data: &[u8], index: usize) -> Result<u8, UsbMidiEventPacketError> {
     match data.get(index) {
         Some(byte) => Ok(*byte),
@@ -212,6 +356,7 @@ fn get_byte_at_position(data: &[u8], index: usize) -> Result<u8, UsbMidiEventPac
 mod tests {
     use crate::message::channel::Channel::{Channel1, Channel2};
     use crate::message::control_function::ControlFunction;
+    use crate::message::data::u14::U14;
     use crate::message::data::u7::U7;
     use crate::message::notes::Note;
     use crate::message::Message;
@@ -237,7 +382,18 @@ mod tests {
         polyphonic_aftertouch: ([10, 160, 36, 64], Message::PolyphonicAftertouch(Channel1, Note::C2, U7(64)).into_packet(Cable0)),
         program_change: ([28, 192, 127, 0], Message::ProgramChange(Channel1, U7(127)).into_packet(Cable1)),
         channel_aftertouch: ([13, 208, 127, 0], Message::ChannelAftertouch(Channel1, U7(127)).into_packet(Cable0)),
-        pitch_wheel: ([14, 224, 64, 32], Message::PitchWheelChange(Channel1, U7(64), U7(32)).into_packet(Cable0)),
+        pitch_wheel: ([14, 224, 64, 32], Message::PitchWheelChange(Channel1, U14(4160)).into_packet(Cable0)),
         control_change: ([11, 177, 1, 32], Message::ControlChange(Channel2, ControlFunction::MOD_WHEEL_1, U7(32)).into_packet(Cable0)),
+        mtc_quarter_frame: ([0x02, 0xF1, 12, 0], Message::MtcQuarterFrame(U7(12)).into_packet(Cable0)),
+        song_position_pointer: ([0x03, 0xF2, 38, 75], Message::SongPositionPointer(U14(9638)).into_packet(Cable0)),
+        song_select: ([0x02, 0xF3, 4, 0], Message::SongSelect(U7(4)).into_packet(Cable0)),
+        tune_request: ([0x05, 0xF6, 0, 0], Message::TuneRequest.into_packet(Cable0)),
+        timing_clock: ([0x0F, 0xF8, 0, 0], Message::TimingClock.into_packet(Cable0)),
+        tick: ([0x0F, 0xF9, 0, 0], Message::Tick.into_packet(Cable0)),
+        start: ([0x0F, 0xFA, 0, 0], Message::Start.into_packet(Cable0)),
+        continue_: ([0x0F, 0xFB, 0, 0], Message::Continue.into_packet(Cable0)),
+        stop: ([0x0F, 0xFC, 0, 0], Message::Stop.into_packet(Cable0)),
+        active_sensing: ([0x0F, 0xFE, 0, 0], Message::ActiveSensing.into_packet(Cable0)),
+        reset: ([0x0F, 0xFF, 0, 0], Message::Reset.into_packet(Cable0)),
     }
 }